Add an overload for Pandora for the member function propagateToLayer
This overload is used by Pandora for propagating tracks without fitting. There
exists another overload that takes an EDM4hep hit and we could pass the last hit
that was added to the track to it. However, that doesn't work because that
overload will try to get the TKalTrackSite (track measurement) that is only
built after fitting, which Pandora does not need since it uses the refitted
tracks (for CLD).